Add initial rsync run
This adds support for an initial rsync run just after all local paths
are watched by inotify. The pyinotify events are handled already and
changed items are recorded in the queue while the initial rsync is
running. The worker blocks until the forked rsync process has finished
and will continue to work with the queue as usual afterwards.

Note that a failing initial sync to a server will prevent any further
activity for that server. The server's worker thread returns and a log
entry will be written indicating the removal of the failing destination.