Add worker queue optimization and keep failed items
This change will keep items that failed to sync in the worker queue
forever and logs a error message. The old behaviour was to discard
failed items which can lead to lots of changes that are not replicated
until the next change happens in the same path.

In addition, the worker queue is optimized to start with least specific
items first and deleted items are not synchronized anymore.

If we delete a tree like this:

    /foo/bar/baz

Then we will record events for /foo/bar, /foo and / in order. However,
since /foo and all its subdirs are already deleted, we can skip all of
them and a single rsync of / will catch them all.

Since there is no simple way to do that in the event handler (we receive
a deletion event for /foo/bar/baz and would have to store this forever
until the deletion of the parent happens?) we do this by stat()ing each
item in the worker.

Suppose we receive dir creation events in this order:

    /foo
    /foo/bar
    /foo/baz

Starting the non-recursive sync with /foo first will enable one rsync
run to create all missing subdirs at once. The old behaviour creates a
large chain of parent dirs per rsync run in the worst case.