Add Item and ItemQueue classes and optimize synchronization
This is the base change for future queue optimizations with support for
the recursive synchronization of items.

For example, if a directory is moved we don't sync each and every subdir
within this renamed dir anymore. We collect all the changes as usual but
optimize the changeset prior to processing it. This means we can now
get rid of all items that are below the path of the moved directory
because they are all handled by one recursive rsync invocation.

After all, this is a huge performance benefit.