Optimize synchronization of newly created directories
This is the same optimization as for directory renaming and gets rid of
lots of slow non-recursive rsync invocations. Note that we're unable to
detect if a whole new subtree was created or just one single dir. In the
latter case there will be no performance impact, but the former case
will see a huge improvement.