Implement classes Source and Destination, add virtual root support
This patch adds support for the rsync-like virtual root marker within
the source path definition. Now you are able to limit the amount of
path information that is sent to the receiving side.

Example:

    pylsyncd /foo/./bar/baz remote:/tmp

Will result in these dirs (including contents):

    /tmp/bar
    /tmp/bar/baz

on the remote side. The local /foo is dropped, because it is the
virtual root for remote paths.