Require at least pyinotify version 0.8.8
Pyinotify version 0.8.8 is the first upstream version with two critical
issues fixed. In addition the 0.8 API is not fully compatible with earlier
versions. We don't support pyinotify 0.7 because we would have to
implement logic that belongs into pyinotify and is much more efficient
there. These are the two important issues:

http://trac.dbzteam.org/pyinotify/ticket/20
http://trac.dbzteam.org/pyinotify/ticket/31