Implement support for rsync file-lists to reduce forking
This change is a *HUGE* performance improvement. Instead of forking one
rsync process for each and every changed source path, we assemble a
rsync-compatible file-list and hand it over to rsync on stdin.

In addition we don't need to assemble source paths containing the
virtual root marker anymore. We let rsync handle this, based on the fact
that all paths within a file-list are relative to the source given on
the command line.

Since we want to optimize the synchronization and need to distinguish
between recursive and non-recursive rsync invocations, the final changes
are:

    1. ItemQueue now has attributes called "dirs" and "trees".

        These are two lists to hold non-recursive dir paths
        ("dirs") and recursive dir paths ("trees"). We prevent
        duplicates and try to optimize out as much as possible.

    2. After ItemQueue optimization the rsync file-lists are assembled

        These rsync file-lists are \0 delimited to support all possible
        pathnames. Also, all paths in the file-list are relative to the
        given rsync source path. The old virtual root ("target path
        rewriting") behaviour is preserved, ie. the full path is
        preserved on the target (default) -- or, if you specify a
        virtual root within the source path, all target paths will be
        relative to this virtual root.

    3. At most two rsync invocations per cycle are done now

        One running recursively with all "trees" and the other running
        non-recursively with all "dirs". If one of both variants has an
        empty queue, we silently skip it.

The old "failsafe" behaviour is preserved, ie. if a destination fails to
sync, the queued items won't be dropped and will be handled by the next
run... (until we hit a limit and remove the destination, as always).